web-animations/animation-player/animation-player-constructor.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer constructor tests</title>
<meta name="assert" content="Creates a new AnimationPlayer object">
<link rel="help" href="http://w3c.github.io/web-animations/#the-animationplayer-interface">
<link rel="help" href="http://w3c.github.io/web-animations/#set-the-timeline-of-a-player">
<link rel="help" href="http://w3c.github.io/web-animations/#set-the-source-content-of-a-player">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = new AnimationPlayer();

    assert_equals(player.source, null, 'If constructor source parameter is ommitted ' +
        'then source attribute of newly created player should be null');
    assert_equals(player.timeline, null, 'If constructor timeline parameter is ommitted ' +
        'then timeline attribute of newly created player should be null');
}, 'The AnimationPlayer constructor without parameters creates ' +
    'AnimationPlayer instance with null source and null timeline');

test(function() {
    var player = new AnimationPlayer(null);

    assert_equals(player.source, null, 'If constructor source parameter is null ' +
        'then source attribute of newly created player should be null');
    assert_equals(player.timeline, null, 'If constructor timeline parameter is ommitted ' +
        'then timeline attribute of newly created player should be null');
}, 'The AnimationPlayer constructor with null source and no timeline creates ' +
    'AnimationPlayer instance with null source and null timeline');

test(function() {
    var player = new AnimationPlayer(null, null);

    assert_equals(player.source, null, 'If constructor source parameter is null ' +
        'then source attribute of newly created player should be null');
    assert_equals(player.timeline, null, 'If constructor timeline parameter is null ' +
        'then timeline attribute of newly created player should be null');
}, 'The AnimationPlayer constructor with null source and timeline parameters creates ' +
    'AnimationPlayer instance with null source and null timeline');

test(function() {
    var sources = [
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION),
        new AnimationGroup([]),
        new AnimationSequence([])
    ];
    sources.forEach(function(source) {
        var player = new AnimationPlayer(source);

        assert_equals(player.source, source, 'Player should be associated ' +
            'to source for source type ' + type(source));
    });
}, 'The AnimationPlayer constructor with source parameter creates ' +
    'AnimationPlayer instance with provided source');

test(function() {
    var timeline = document.timeline;
    var player = new AnimationPlayer(null, timeline);

    assert_equals(player.timeline, timeline, 'Timeline of newly created player ' +
        'should be specified by constructor timeline parameter');
}, 'The AnimationPlayer constructor with timeline parameter creates ' +
    'AnimationPlayer instance with provided timeline');

test(function() {
    var nodes = [
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION),
        new AnimationGroup([]),
        new AnimationSequence([])
    ];
    var parent = new AnimationGroup(nodes);

    nodes.forEach(function(node){
        var player = new AnimationPlayer(node);

        assert_equals(player.source, node, 'Source of newly created player ' +
            'should be specified by constructor source parameter');
        assert_equals(node.parent, null, 'Source should be removed from its parent ' +
            'animation group');
    });
    assert_array_equals(parent.children, [], 'Source should be removed from the parent ' +
        'group\'s children');
}, 'Test AnimationPlayer constructor removes the source content node from its parent group, ' +
    'if any');

test(function() {
    var nodes = [
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION),
        new AnimationGroup([]),
        new AnimationSequence([])
    ];
    var parent = new AnimationSequence(nodes);

    nodes.forEach(function(node){
        var player = new AnimationPlayer(node);

        assert_equals(player.source, node, 'Source of newly created player ' +
            'should be specified by constructor source parameter');
        assert_equals(node.parent, null, 'Source should be removed from its parent ' +
            'animation sequence');
    });
    assert_array_equals(parent.children, [], 'Source should be removed from the parent ' +
        'sequence\'s children');
}, 'Test AnimationPlayer constructor removes the source content node from its parent ' +
    'animation sequence, if any');

test(function() {
    var sources = [
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION),
        new AnimationGroup([]),
        new AnimationSequence([])
    ];
    sources.forEach(function(source) {
        var player1 = new AnimationPlayer(source);
        assert_equals(player1.source, source, 'First player should be associated ' +
            'to source for source type ' + type(source));

        var player2 = new AnimationPlayer(source);
        assert_equals(player1.source, null, 'Previous player should be ' +
            'disassociated from source type ' + type(source));
        assert_equals(player2.source, source, 'Second player should be associated ' +
            'to source for source type ' + type(source));
    });

}, 'Test AnimationPlayer constructor sets source to null in player associated ' +
    'with passed animateion node');
</script>
</body>
